Redirect RSVP'd users that land on /learn or /volunteer to the event show page
Previously, if you had already RSVP'd to an event, then logged out,
clicked the Learn or Volunteer buttons on the event show page, and
logged in with the modal, you would be led to an RSVP
form that wouldn't actually work.

Now, a user in this situation is just redirected to the event show page
where they can decide for themselves their next course of action
(maybe editing their RSVP?).