Use <img> tags to hold SVGs rather than dumping them into the document
The only reason to put them directly in the document was to make them
style-able with CSS, but it had the consequence of not allowing the SVG
images to be cached, and looking real bad on 'view source'.

So I moved the logo text coloring into the SVG and now we can all live
cachily ever after.