Don't display an extra 'true' flash on session timeout.
Previously, when your Devise session had timed out, you would
see two flash messages, one with the actual error and one
that just said 'true'.

This is because Devise sets a key in the flash called :timedout,
which you need to avoid displaying in the view. This was fixed
before, but in recent versions of Rails (4.1?) flash keys are
normalized to strings instead of symbols, so it showed up again.