Validate at least one session is selected for an upcoming attendee RSVP
An RSVP that has no sessions selected should probably just be cancelled!

This only applies to upcoming events because there are a lot of
records in the production DB that have no associated sessions.

The validation message for this is not great yet ("Rsvp sessions
must be selected")

The RSVP factory now accepts a transient attribute, a hash of
'session_checkins'. This removes most of the use cases for the
rsvp_sessions factory. If 'session_checkins' are not provided,
an rsvp_sessions record will be made for the first of the
event_sessions.