Refactor user index page to reduce # of queries / memory usage
Previously, since the .meetup_id for regular users was moved to
the Authentications table, the page needed to make an
Authentications query for every user that it rendered. Now,
it just makes one big query up front to build a hash of IDs.

It still locks up the browser during render pretty seriously.
We probably have too many users to realistically render in one
DataTable. But this is a cool bandaid for now.