Add class level title details to the attendee modal.
On the section organizer page, a student's details can be viewed via a
modal. This modal does not currently contain the class level title that
the student has signed up for. In this commit, I have added the
appropriate details to the student's attendee modal. This was achieved
through adding a #levels method to the Event model, which returns all
the levels for a given course. The Rsvp model leverages the Event
model's knowledge to "pick out" a specific level based on the
information provided by the student via a #level method. Then, since we
want the title of the level, I've added a #level_title method that
invokes #level and returns that hash's title. The only other major
change is to attendee_detail.hbs template, which now includes a section
for displaying the student's class level title.