Locations can be set on individual sessions
These locations are considered an 'override' of the event-wide
location, which still exists.

This adds only the editing functionality, later commits will
address showing the location on the event page and in emails.

Also extracts a concern, 'PresenceTrackingBoolean', to aid in adding
checkbox togglers for optional fields in forms without adding an
extra boolean in the database. Currently used in two places:

location_overridden for EventSession#location_id in this commit
needs_childcare for Rsvp#childcare_info previously