Sort 'past events' on the frontpage properly in Safari private mode
The DataTables code wasn't getting a chance to run because
of earlier localStorage exceptions from the Chapter filter.

Safari returns a 'localStorage' object but it raises
QuotaExceededError if you try to set any keys. Everything
that needs to add to localStorage has to check for this
condition in a try/catch :'(