Extract partials for publishing buttons on unpublished_events#index and event#show
Also moves the generation of the 'Publish' confirm text from
a "data: confirm" to a bespoke AJAX request. This allows the
controller to not care about fetching the # of users in each
region just to show a confirm dialog.

To facilitate this, added a /regions/:id.json which can
provide the name and # of users to be emailed for a specific region.