De-flake the CoC URL test
Now that the CoC URL is retrieved asynchronously, it takes
a little bit of time for the CoC url to change after the
chapter has changed. Though a user would probably never see
this, Capybara of course manages to, so it now polls the
href until it looks like the right one.