Fix Course->Level validations to allow more flexible reordering
Previously, if you have two levels originally positioned as 1 and 2,
and swapped their indices in the editor, it would never save because
each level would try to save itself and hit the per-level uniqueness
validation.

Now, the validation is handled on the Course, so you can swap the
level positions as much as you want as long as they all end up
unique in the end.

Also added a bunch of not-great javascript to swap the order of
fields on the Course editor as Level positions are changed.